+++ src/Firebase/Auth/CreateActionLink.php
@@ -10,20 +10,25 @@ use Kreait\Firebase\Value\Email;
 final class CreateActionLink
 {
     private string $type;
-    private Email $email;
+    private string $email;
     private ActionCodeSettings $settings;
     private ?string $tenantId = null;
     private ?string $locale = null;
 
-    private function __construct(string $type, Email $email, ActionCodeSettings $settings)
+    private function __construct(string $type, string $email, ActionCodeSettings $settings)
     {
         $this->type = $type;
         $this->email = $email;
         $this->settings = $settings;
     }
 
-    public static function new(string $type, Email $email, ActionCodeSettings $settings, ?string $tenantId = null, ?string $locale = null): self
+    /**
+     * @param \Stringable|string $email
+     */
+    public static function new(string $type, $email, ActionCodeSettings $settings, ?string $tenantId = null, ?string $locale = null): self
     {
+        $email = (string) (new Email((string) $email));
+
         $instance = new self($type, $email, $settings);
         $instance->tenantId = $tenantId;
         $instance->locale = $locale;
@@ -36,7 +41,7 @@ final class CreateActionLink
         return $this->type;
     }
 
-    public function email(): Email
+    public function email(): string
     {
         return $this->email;
     }