+++ src/Firebase/Auth/SendActionLink.php
@@ -23,7 +23,7 @@ final class SendActionLink
         return $this->action->type();
     }
 
-    public function email(): Email
+    public function email(): string
     {
         return $this->action->email();
     }