+++ src/Firebase/Factory.php
@@ -75,7 +75,7 @@ class Factory
 
     protected ?ProjectId $projectId = null;
 
-    protected ?Email $clientEmail = null;
+    protected ?string $clientEmail = null;
 
     protected CacheInterface $verifierCache;
 
@@ -137,7 +137,7 @@ class Factory
     public function withClientEmail(string $clientEmail): self
     {
         $factory = clone $this;
-        $factory->clientEmail = new Email($clientEmail);
+        $factory->clientEmail = (string) (new Email($clientEmail));
 
         return $factory;
     }
@@ -316,7 +316,7 @@ class Factory
         throw new RuntimeException('Unable to determine the Firebase Project ID');
     }
 
-    protected function getClientEmail(): ?Email
+    protected function getClientEmail(): ?string
     {
         if ($this->clientEmail !== null) {
             return $this->clientEmail;
@@ -325,7 +325,7 @@ class Factory
         $serviceAccount = $this->getServiceAccount();
 
         if ($serviceAccount !== null) {
-            return $this->clientEmail = new Email($serviceAccount->getClientEmail());
+            return $this->clientEmail = (string) (new Email($serviceAccount->getClientEmail()));
         }
 
         if ($this->discoveryIsDisabled) {
@@ -338,7 +338,7 @@ class Factory
                 && ($credentials instanceof SignBlobInterface)
                 && ($clientEmail = $credentials->getClientName())
             ) {
-                return $this->clientEmail = new Email($clientEmail);
+                return $this->clientEmail = $clientEmail;
             }
         } catch (Throwable $e) {
             return null;