+++ CHANGELOG.md
@@ -5,8 +5,13 @@
 * It is now mandatory to provide a Firebase Project ID. When the project ID cannot be determined from the
   provided credentials (usually a service account), it can be provided by setting the `GOOGLE_CLOUD_PROJECT=<project-id>`
   environment variable.
+* The following methods now return strings instead of value objects:
+    * `Kreait\Firebase\Contract\Auth::confirmPasswordResetAndReturnEmail()`
+    * `Kreait\Firebase\Contract\Auth::verifyPasswordResetCodeAndReturnEmail()`
+    * `Kreait\Firebase\RemoteConfig\User::email()`
 * The following classes are used for validation and have been marked internal. They shouldn't be used directly, as they
   could be updated with breaking changes or get removed entirely in the future.
+    * `Kreait\Firebase\Value\Email`
     * `Kreait\Firebase\Value\Uid`
     * `Kreait\Firebase\Value\Url`